LayoutTests/http/tests/misc/css-reject-any-type-in-strict-mode.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0//EN">
<html>
<head>
    <title></title>
    <link rel="stylesheet" href="resources/stylesheet.php">
    <script>
        function test()
        {
            if (window.testRunner)
                testRunner.dumpAsText();
            
            var target = document.getElementById("target");
            var style = getComputedStyle(target);
            target.innerText = style.position == "static" ? "SUCCESS" : "FAIL";
        }
    </script>
</head>
<body onload="test()">
    <p>
        Related test to <i><a href="http://bugs.webkit.org/show_bug.cgi?id=11451">http://bugs.webkit.org/show_bug.cgi?id=11451</a>
        REGRESSION: Dell.com does not render correctly in ToT (stylesheet not loaded)</i>.  Making sure that in strict mode we would not load such a stylesheet.
    </p>
    <p id="target">
    </p>
</body>
</html>
